Project6_Nested_Serializer/requirements.txt
Django>=5.2
djangorestframework==3.15.2
asgiref==3.8.1
sqlparse==0.5.0
tzdata==2024.1